# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSE AI Agent - Construction Site Safety Monitoring API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SiteStatus = typeof SiteStatus[keyof typeof SiteStatus];


export const SiteStatus = {
  active: 'active',
  paused: 'paused',
  offline: 'offline',
} as const;

export interface Site {
  id: number;
  /** Arabic site name */
  nameAr: string;
  nameEn?: string;
  location: string;
  /** Compliance percentage 0-100 */
  complianceRate: number;
  status: SiteStatus;
  cameraCount: number;
  /** @nullable */
  lastInspection?: string | null;
}

export type ViolationViolationType = typeof ViolationViolationType[keyof typeof ViolationViolationType];


export const ViolationViolationType = {
  no_helmet: 'no_helmet',
  no_vest: 'no_vest',
  smoking_danger_zone: 'smoking_danger_zone',
  blocked_exit: 'blocked_exit',
  no_harness: 'no_harness',
  other: 'other',
} as const;

export type ViolationStatus = typeof ViolationStatus[keyof typeof ViolationStatus];


export const ViolationStatus = {
  open: 'open',
  resolved: 'resolved',
} as const;

export interface Violation {
  id: number;
  siteId: number;
  ticketNumber: string;
  violationType: ViolationViolationType;
  violationTypeAr?: string;
  zone: string;
  detectedAt: string;
  /** @nullable */
  resolvedAt?: string | null;
  potentialFineAmount: number;
  status: ViolationStatus;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type ViolationInputViolationType = typeof ViolationInputViolationType[keyof typeof ViolationInputViolationType];


export const ViolationInputViolationType = {
  no_helmet: 'no_helmet',
  no_vest: 'no_vest',
  smoking_danger_zone: 'smoking_danger_zone',
  blocked_exit: 'blocked_exit',
  no_harness: 'no_harness',
  other: 'other',
} as const;

export interface ViolationInput {
  siteId: number;
  violationType: ViolationInputViolationType;
  zone: string;
  potentialFineAmount: number;
  notes?: string;
}

export type ViolationUpdateStatus = typeof ViolationUpdateStatus[keyof typeof ViolationUpdateStatus];


export const ViolationUpdateStatus = {
  open: 'open',
  resolved: 'resolved',
} as const;

export interface ViolationUpdate {
  status?: ViolationUpdateStatus;
  notes?: string;
  resolvedAt?: string;
}

export interface DashboardStats {
  complianceRate: number;
  complianceWeeklyDelta: number;
  totalFinesAvoided: number;
  activeTickets: number;
  avgResolutionMinutes: number;
  monitoredWorkers: number;
  activeAlerts: number;
}

export interface CompliancePoint {
  week: string;
  rate: number;
}

export interface ViolationBreakdownItem {
  type: string;
  typeAr: string;
  count: number;
  percentage: number;
}

export type RecommendationPriority = typeof RecommendationPriority[keyof typeof RecommendationPriority];


export const RecommendationPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Recommendation {
  id: number;
  siteId: number;
  priority: RecommendationPriority;
  titleAr: string;
  descriptionAr: string;
  expectedSaving: number;
  timeWindow: string;
}

export type MonitoringSessionInputSource = typeof MonitoringSessionInputSource[keyof typeof MonitoringSessionInputSource];


export const MonitoringSessionInputSource = {
  cctv: 'cctv',
  mobile: 'mobile',
} as const;

export interface MonitoringSessionInput {
  siteId: number;
  source: MonitoringSessionInputSource;
  detectionCategories?: string[];
}

export interface MonitoringResult {
  sessionId: string;
  siteId: number;
  detectedViolations: Violation[];
  newComplianceRate: number;
  alertSent: boolean;
  /** @nullable */
  alertMessage?: string | null;
}

export type ListViolationsParams = {
siteId?: number;
status?: ListViolationsStatus;
};

export type ListViolationsStatus = typeof ListViolationsStatus[keyof typeof ListViolationsStatus];


export const ListViolationsStatus = {
  open: 'open',
  resolved: 'resolved',
  all: 'all',
} as const;

export type GetDashboardStatsParams = {
siteId?: number;
};

export type GetComplianceHistoryParams = {
siteId?: number;
};

export type GetViolationBreakdownParams = {
siteId?: number;
};

export type ListRecommendationsParams = {
siteId?: number;
};