# lib/api-zod/src/generated/types/compliancePoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSE AI Agent - Construction Site Safety Monitoring API
 * OpenAPI spec version: 0.1.0
 */

export interface CompliancePoint {
  week: string;
  rate: number;
}