# lib/api-zod/src/generated/types/recommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSE AI Agent - Construction Site Safety Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { RecommendationPriority } from './recommendationPriority';

export interface Recommendation {
  id: number;
  siteId: number;
  priority: RecommendationPriority;
  titleAr: string;
  descriptionAr: string;
  expectedSaving: number;
  timeWindow: string;
}