# lib/api-zod/src/generated/types/violationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HSE AI Agent - Construction Site Safety Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import type { ViolationInputViolationType } from './violationInputViolationType';

export interface ViolationInput {
  siteId: number;
  violationType: ViolationInputViolationType;
  zone: string;
  potentialFineAmount: number;
  notes?: string;
}